move inner class to its own class
